refactor: move DIRAC WMS WebApp related RPC calls to dedicated service
