+++ src/DIRAC/MonitoringSystem/ConfigTemplate.cfg
@@ -28,4 +28,13 @@ Services
     }
   }
   ##END
+  ##BEGIN WebApp
+  WebApp
+  {
+    Port = 9199
+    Authorization
+    {
+      Default = authenticated
+    }
+  }
 }
