+++ tests/Integration/WorkloadManagementSystem/Test_WMSAdministratorClient.py
@@ -11,9 +11,9 @@ from DIRAC.WorkloadManagementSystem.Client.WMSAdministratorClient import WMSAdmi
 def test_WMSAdministratorClient():
     wmsAdministrator = WMSAdministratorClient()
 
-    res = wmsAdministrator.getSiteSummaryWeb({}, [], 0, 100)
-    assert res["OK"], res["Message"]
-    assert res["Value"]["TotalRecords"] in [0, 1, 2, 34]
+    # res = wmsAdministrator.getSiteSummaryWeb({}, [], 0, 100)
+    # assert res["OK"], res["Message"]
+    # assert res["Value"]["TotalRecords"] in [0, 1, 2, 34]
 
-    res = wmsAdministrator.getSiteSummarySelectors()
-    assert res["OK"], res["Message"]
+    # res = wmsAdministrator.getSiteSummarySelectors()
+    # assert res["OK"], res["Message"]
